cargo bench: extract startup bench into bench-helper
