+++ benchmark/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod cpio;
 pub mod terminal;
 pub mod vm;
+pub mod startup_base;
